% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmer.R
\name{.split_data}
\alias{.split_data}
\title{_split_data}
\usage{
.split_data(data, group)
}
\arguments{
\item{data}{data.table to split}

\item{group}{\code{data} to be splitted by the \code{group} column}
}
\value{
list
}
\description{
_split_data
}
