% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmer.R
\name{plmer}
\alias{plmer}
\title{lmer}
\usage{
plmer(dat, group, formula, ncore = 4, verbose = TRUE)
}
\arguments{
\item{group}{perform lmer for \code{data} grouped by the \code{group} columns}

\item{formula}{formula for lmer function}

\item{ncore}{number of cores to use}

\item{verbose}{logical}

\item{data}{data.table to perform}
}
\value{
data.table
}
\description{
run lmer in parallel
}
\examples{
system.time(ret <- plmer(dat[c("A1BG", "AAMP")], "genes", value~(1|tissues) + (1|individual), ncore = 16))
}
